[#8] modify: Cafe 도메인 관련 테스트 수정
- p6spy 모듈 제거
- CafeRepositoryTest BaseTimeEntity auditing 이슈 해결
- CafeTestUtils로 공통 함수 관리
- application.yml 설정 변경(주로 logging 관련)
