+++ .gitignore
@@ -34,6 +34,3 @@ out/
 
 ### VS Code ###
 .vscode/
-
-### log ###
-*.log
