+++ src/test/resources/application.yml
@@ -1,8 +1,13 @@
 spring:
   flyway:
     enabled: false
+  jpa:
+    properties:
+      hibernate:
+        format_sql: true
+        show_sql: false
 
 logging:
   level:
-    org.hibernate.SQL: debug # logger 통해 로깅
+    org.hibernate.SQL: debug
     org.hibernate.type: trace
